+++ client/src/styles/Admin/Paginated.css
